--- framework/Sources/VGSFramework/UIElements/Text Field/Mask/MaskedTextField+padding.swift
@@ -1,23 +1,0 @@
-//
-//  MaskedTextField+padding.swift
-//  VGSFramework
-//
-//  Created by Vitalii Obertynskyi on 9/28/19.
-//  Copyright © 2019 Vitalii Obertynskyi. All rights reserved.
-//
-
-import Foundation
-
-public extension MaskedTextField {    
-    override open func textRect(forBounds bounds: CGRect) -> CGRect {
-        return bounds.inset(by: padding)
-    }
-
-    override open func placeholderRect(forBounds bounds: CGRect) -> CGRect {
-        return bounds.inset(by: padding)
-    }
-
-    override open func editingRect(forBounds bounds: CGRect) -> CGRect {
-        return bounds.inset(by: padding)
-    }
-}
